Updated calendar.js
Added a new function called jumpTo to allow for navigating to a month and year directly. For my purposes, I have a month and year dropdown and did not want to use the next and prev buttons.

Also added popover functionality to allow for the use of popovers on the events instead of tooltips. I needed more details to be displayed in the tooltip of the event on the month view so I decided to use a popover.